chore: UtxoSecret and split out addr
