+++ crates/mithras-crypto/src/lib.rs
@@ -1,3 +1,6 @@
+pub mod address;
+
+use crate::address::MithrasAddr;
 use anyhow::{Result, anyhow};
 use base64::{Engine as _, engine::general_purpose::STANDARD_NO_PAD as B64};
 use curve25519_dalek::scalar::clamp_integer;
@@ -13,69 +16,6 @@ use rand::rngs::OsRng;
 use serde::{Deserialize, Serialize};
 use sha2::{Digest, Sha256, Sha512};
 use x25519_dalek::{PublicKey as X25519PublicKey, StaticSecret};
-
-use bech32::{self, Hrp};
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct MithrasAddr {
-    pub version: u8,
-    pub network: u8,
-    pub suite: u8,
-    pub spend_ed25519: [u8; 32],
-    pub disc_x25519: [u8; 32],
-}
-
-impl MithrasAddr {
-    pub fn encode(&self) -> String {
-        let mut data = Vec::<u8>::with_capacity(3 + 32 + 32);
-        data.push(self.version);
-        data.push(self.network);
-        data.push(self.suite);
-        data.extend_from_slice(&self.spend_ed25519);
-        data.extend_from_slice(&self.disc_x25519);
-        bech32::encode::<bech32::Bech32m>(Hrp::parse_unchecked("mith"), &data).unwrap()
-    }
-
-    pub fn decode(s: &str) -> Result<Self> {
-        let (hrp, data) = bech32::decode(s)?;
-        if hrp.as_str() != "mith" {
-            return Err(anyhow!(
-                "invalid human-readable prefix. Got {}, expected mith",
-                hrp
-            ));
-        }
-        let version = data[0];
-        let network = data[1];
-        let suite = data[2];
-        let spend = data[3..35].try_into()?;
-        let disc = data[35..67].try_into()?;
-        Ok(Self {
-            version,
-            network,
-            suite,
-            spend_ed25519: spend,
-            disc_x25519: disc,
-        })
-    }
-
-    pub fn from_keys(
-        spend: &VerifyingKey,
-        disc: &X25519PublicKey,
-        version: u8,
-        network: u8,
-        suite: u8,
-    ) -> Self {
-        let spend_bytes = spend.to_bytes();
-        let disc_bytes = disc.as_bytes();
-        Self {
-            version,
-            network,
-            suite,
-            spend_ed25519: spend_bytes,
-            disc_x25519: *disc_bytes,
-        }
-    }
-}
 
 pub fn suite() -> Hpke<HpkeLibcrux> {
     Hpke::new(
@@ -96,13 +36,13 @@ pub struct HpkeEnvelope {
     pub ciphertext_b64: String,
 }
 
-pub struct MithrasSecret {
+pub struct UtxoSecrets {
     pub spending_secret: [u8; 32],
     pub nullifier_secret: [u8; 32],
     pub amount: u64,
 }
 
-impl From<[u8; 72]> for MithrasSecret {
+impl From<[u8; 72]> for UtxoSecrets {
     fn from(bytes: [u8; 72]) -> Self {
         let mut spending_secret = [0u8; 32];
         let mut nullifier_secret = [0u8; 32];
@@ -122,8 +62,8 @@ impl From<[u8; 72]> for MithrasSecret {
     }
 }
 
-impl From<MithrasSecret> for [u8; 72] {
-    fn from(secret: MithrasSecret) -> Self {
+impl From<UtxoSecrets> for [u8; 72] {
+    fn from(secret: UtxoSecrets) -> Self {
         let mut bytes = [0u8; 72];
         bytes[0..32].copy_from_slice(&secret.spending_secret);
         bytes[32..64].copy_from_slice(&secret.nullifier_secret);
@@ -431,7 +371,7 @@ mod tests {
             .setup_sender(hpke_recipient.public_key(), info, None, None, None)
             .unwrap();
 
-        let mithras_secret = MithrasSecret {
+        let mithras_secret = UtxoSecrets {
             spending_secret: [42u8; 32],
             nullifier_secret: [43u8; 32],
             amount: 1000,
@@ -561,7 +501,7 @@ mod tests {
             .setup_sender(hpke_recipient.public_key(), info, None, None, None)
             .unwrap();
 
-        let mithras_secret = MithrasSecret {
+        let mithras_secret = UtxoSecrets {
             spending_secret: [42u8; 32],
             nullifier_secret: [43u8; 32],
             amount: 1000,
@@ -615,4 +555,3 @@ mod tests {
         Ok(())
     }
 }
-
